+++ PKHeX.Core/Legality/Encounters/EncounterStatic/EncounterStatic.cs
@@ -268,7 +268,7 @@ public abstract record EncounterStatic(GameVersion Version) : IEncounterable, IM
     {
         if (IsRandomUnspecificForm)
             return true;
-        return Form == evo.Form || FormInfo.IsFormChangeable((ushort)Species, Form, pk.Form, pk.Format);
+        return Form == evo.Form || FormInfo.IsFormChangeable(Species, Form, pk.Form, pk.Format);
     }
 
     // override me if the encounter type has any eggs
