+++ PKHeX.Core/Legality/Verifiers/MiscVerifier.cs
@@ -470,7 +470,7 @@ public sealed class MiscVerifier : Verifier
         if (pk.Species != data.EncounterMatch.Species)
             return; // evolved
 
-        if (Unfeedable.Contains((ushort)pk.Species))
+        if (Unfeedable.Contains(pk.Species))
             data.AddLine(GetInvalid(string.Format(LMemoryStatFullness, "0"), Encounter));
     }
