+++ PKHeX.Core/Saves/Substructures/Gen3/SecretBase3Team.cs
@@ -52,7 +52,7 @@ public sealed class SecretBase3Team
     {
         var pk = Team[index];
         WriteUInt32LittleEndian(Data.AsSpan(GetOffsetPID(index)), pk.PID);
-        WriteUInt16LittleEndian(Data.AsSpan(GetOffsetSpecies(index)), (ushort)pk.Species);
+        WriteUInt16LittleEndian(Data.AsSpan(GetOffsetSpecies(index)), pk.Species);
         WriteUInt16LittleEndian(Data.AsSpan(GetOffsetItem(index)), (ushort)pk.HeldItem);
         WriteUInt16LittleEndian(Data.AsSpan(GetOffsetMove(index, 0)), (ushort)pk.Move1);
         WriteUInt16LittleEndian(Data.AsSpan(GetOffsetMove(index, 1)), (ushort)pk.Move2);
