+++ PKHeX.Drawing.PokeSprite/Util/SpriteName.cs
@@ -71,7 +71,7 @@ public static class SpriteName
     /// <summary>
     /// Species that show their default Species sprite regardless of current <see cref="PKM.Form"/>
     /// </summary>
-    private static readonly HashSet<int> SpeciesDefaultFormSprite = new()
+    private static readonly HashSet<ushort> SpeciesDefaultFormSprite = new()
     {
         (int)Species.Mothim,
         (int)Species.Scatterbug,
@@ -86,7 +86,7 @@ public static class SpriteName
     /// <summary>
     /// Species that show a <see cref="PKM.Gender"/> specific Sprite
     /// </summary>
-    private static readonly HashSet<int> SpeciesGenderedSprite = new()
+    private static readonly HashSet<ushort> SpeciesGenderedSprite = new()
     {
         (int)Species.Pikachu,
         (int)Species.Hippopotas,
